Token objects for Insee applications

Cette classe représente les jetons d'accès aux applications créées sur api.insee.fr et hérite de la classe Token2.0 du package httr. Les objets de cette classe doivent être créés en utilisant le constructeur insee_token().

Format

Un objet de classe R6.

Methods

has_expired() : le jeton d'accès a-t-il expiré ?

cache() : sauvegarde le jeton d'accès dans un cache.

revoke() : révoque le jeton d'accès.

refresh() : rafraichit le jeton d'accès (le point d'accès de rafraichissement OAuth2 n'étant pas disponible, le jeton d'accès courant est révoqué puis un nouveau jeton d'accès est généré).

Caching

OAuth tokens are cached on disk in a file called .httr-oauth saved in the current working directory. Caching is enabled if:

The session is interactive, and the user agrees to it, OR

The .httr-oauth file is already present, OR

getOption("httr_oauth_cache") is TRUE

You can suppress caching by setting the httr_oauth_cache option to FALSE.

Tokens are cached based on their endpoint and parameters.

The cache file should not be included in source code control or R packages (because it contains private information), so httr will automatically add the appropriate entries to .gitignore and .Rbuildignore if needed.

Super classes

httr::Token -> httr::Token2.0 -> TokenInsee

Methods

Public methods

TokenInsee$print()

TokenInsee$init_credentials()

TokenInsee$has_expired()

TokenInsee$load_from_cache()

TokenInsee$revoke()

TokenInsee$refresh()

TokenInsee$can_refresh()

TokenInsee$clone()

Inherited methods

httr::Token$cache()

httr::Token$hash()

httr::Token$initialize()

httr::Token2.0$sign()

httr::Token2.0$validate()

Method print()

Usage

TokenInsee$print(...)

Method init_credentials()

Usage

TokenInsee$init_credentials()

Method has_expired()

Usage

TokenInsee$has_expired()

Method load_from_cache()

Usage

TokenInsee$load_from_cache()

Method revoke()

Usage

TokenInsee$revoke()

Method refresh()

Usage

TokenInsee$refresh()

Method can_refresh()

Usage

TokenInsee$can_refresh()

Method clone()

The objects of this class are cloneable with this method.

Usage

TokenInsee$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.