Addresses of the Insee APIs

Cette fonction renvoie les adresses des API de l'Insee.

insee_scopes(
  api = c("Sirene - V3", "Nomenclatures - v1", "DonneesLocales - V0.1", "BDM - V1"),
  insee_url = getOption("apinsee.url")
)

Arguments

api
Un vecteur de chaînes de caractères dont chaque élément comprend le nom d'une API. La correspondance partielle est acceptée.

insee_url
Adresse du site fournissant les API. Ce paramètre n'est utile que pour un usage interne à l'Insee, voir la section "Utilisation interne à l'Insee".

Value

Un vecteur de chaînes de caractères comprenant les adresses des API de l'Insee.

Utilisation interne à l'Insee

Dans le cas d'utilisation standard, la valeur de l'option apinsee.url est "https://api.insee.fr/". Il n'est pas utile de la modifier sauf pour les agents de l'Insee qui souhaiteraient utiliser les plateformes de test, de recette ou de pré-production. Dans ce cas, afin de rendre les programmes portables d'un environnement à un autre.il est préférable de déclarer l'adresse en modifiant l'option apinsee.url : options(apinsee.url = "<URL>").

Examples

insee_scopes()

#> [1] "https://api.insee.fr/entreprises/sirene/V3"       
#> [2] "https://api.insee.fr/entreprises/sirene"          
#> [3] "https://api.insee.fr/metadonnees/nomenclatures/v1"
#> [4] "https://api.insee.fr/metadonnees/nomenclatures"   
#> [5] "https://api.insee.fr/donnees-locales/V0.1"        
#> [6] "https://api.insee.fr/series/BDM/V1"               
#> [7] "https://api.insee.fr/series/BDM"                  

insee_scopes("Sirene")

#> [1] "https://api.insee.fr/entreprises/sirene/V3"
#> [2] "https://api.insee.fr/entreprises/sirene"   

insee_scopes("Nomenclatures")

#> [1] "https://api.insee.fr/metadonnees/nomenclatures/v1"
#> [2] "https://api.insee.fr/metadonnees/nomenclatures"   

insee_scopes(c("Sirene", "Nomenclatures", "DonneesLocales", "BDM"))

#> [1] "https://api.insee.fr/entreprises/sirene/V3"       
#> [2] "https://api.insee.fr/entreprises/sirene"          
#> [3] "https://api.insee.fr/metadonnees/nomenclatures/v1"
#> [4] "https://api.insee.fr/metadonnees/nomenclatures"   
#> [5] "https://api.insee.fr/donnees-locales/V0.1"        
#> [6] "https://api.insee.fr/series/BDM/V1"               
#> [7] "https://api.insee.fr/series/BDM"