Generate a valid token for an Insee application

Cette fonction génère un jeton d'accès valide à une application créée sur le site api.insee.fr.

insee_token(
  app,
  cache = getOption("httr_oauth_cache"),
  config_init = list(),
  credentials = NULL,
  validity_period = 604800,
  insee_url = getOption("apinsee.url"),
  api = c("Sirene - V3", "Nomenclatures - v1", "DonneesLocales - V0.1", "BDM - V1")
)

Arguments

app
An OAuth consumer application, created by oauth_app()

cache
A logical value or a string. TRUE means to cache using the default cache file .httr-oauth, FALSE means don't cache, and NA means to guess using some sensible heuristics. A string means use the specified path as the cache file.

config_init
Additional configuration settings sent to POST(), e.g. user_agent().

credentials
Advanced use only: allows you to completely customise token generation.

validity_period
Un entier; durée de validité du jeton d'accès. Cette valeur n'est utilisée que lorsque le dernier jeton d'accès a expiré ou a été révoqué.

insee_url
Adresse du site fournissant les API. Ce paramètre n'est utile que pour un usage interne à l'Insee, voir la section "Utilisation interne à l'Insee".

api
Un vecteur de chaînes de caractères dont chaque élément comprend le ou les noms des API accessibles par l'application. La correspondance partielle est acceptée.

Value

Un objet de classe TokenInsee.

Utilisation interne à l'Insee

Dans le cas d'utilisation standard, la valeur de l'option apinsee.url est "https://api.insee.fr/". Il n'est pas utile de la modifier sauf pour les agents de l'Insee qui souhaiteraient utiliser les plateformes de test, de recette ou de pré-production. Dans ce cas, afin de rendre les programmes portables d'un environnement à un autre.il est préférable de déclarer l'adresse en modifiant l'option apinsee.url : options(apinsee.url = "<URL>").